backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
bandit==1.8.3
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.2.0
flake8-bandit==4.1.1
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
html5lib==1.1
idna==3.10
isort==6.0.1
lxml==5.4.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.2
pathspec==0.12.1
pbr==6.1.1
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.3.2
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.1
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
setuptools==78.1.0
six==1.17.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stevedore==5.4.1
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
webencodings==0.5.1
